Stop tracking yarn.lock
Child projects will get their version of this when ./fix.sh is run;
templating this out and trying to manage the different variants is not
a winning strategy.

Not quite sure how to manage common security updates for child
projects; might need to add a common script in a future PR if the
process isn't trivial.